fix: memory leak when rejected bullets are not freed
The space associated with the Bullet along with the X resources
allocated for it are rejected by list_append() when the list_len()
is greater than or equal to the specified config.bullet_max.

pollfds() has been altered to return a character string from getline()
via getbulletstr(); this still exhausts the input buffer even when
list_len() is already at config.bullet_max and requires less overhead
and intermediary memory to be freed.

Additionally, in a case where EOF has been received but there are
still bullets remaining in the list (such as `xdanmaku <<< "STEAKY"'),
a usleep() call has been added so the final bullets don't immediately
speed off the screen.

The VERSION is set to when the (previously forgotten) work on this
commit started.