fix: edge case where the last bullet in a list is not properly destroyed
bullet_passed() would return true on the last frame where a bullet
may still be visible on the screen, which results in the bullet being
destroyed and the list length being decremented right before the check

    if (list_len(list))
	XFlush(state.dpy);

is made. In other words, the call to destroy the bullet Window is not
made until the next call to XFlush is made to flush the X server's
request buffer, which in this case would be whenever the next time the
list's length is not 0.

The implemented fix is to keep the bullet in the list until its entire
width surpasses the screen (i.e., its right edge is to the left of the
screen's left edge) in order to prevent the artifacting, and then
additionally to store the list's length prior to the list_iter() loop
and use that to determine whether to call XFlush() so the X resources
can be freed sooner.

https://tronche.com/gui/x/xlib/event-handling/handling-output-buffer.html
https://tronche.com/gui/x/xlib/event-handling/XFlush.html